+++ src/main/java/ay2021s1_cs2103_w16_3/finesse/storage/JsonAdaptedFrequentExpense.java
@@ -50,7 +50,7 @@ public class JsonAdaptedFrequentExpense {
     }
 
     /**
-     * Converts this Jackson-friendly adapted expense object into the model's {@code Expense} object.
+     * Converts this Jackson-friendly adapted expense object into the model's {@code FrequentExpense} object.
      *
      * @throws IllegalValueException if there were any data constraints violated in the adapted frequent expense.
      */
